refactor(api): simplify banners endpoint parameter types and remove unused schemas
