+++ packages/rest-typings/src/v1/banners.ts
@@ -1,51 +1,13 @@
 import type { BannerPlatform, IBanner } from '@rocket.chat/core-typings';
 
-import { ajv } from './Ajv';
-
-type Banners = {
-	platform: BannerPlatform;
-};
-
-const BannersSchema = {
-	type: 'object',
-	properties: {
-		platform: {
-			type: 'string',
-			enum: ['web', 'mobile'],
-		},
-	},
-	required: ['platform'],
-	additionalProperties: false,
-};
-
-export const isBannersProps = ajv.compile<Banners>(BannersSchema);
-
-type BannersDismiss = {
-	bannerId: string;
-};
-
-const BannersDismissSchema = {
-	type: 'object',
-	properties: {
-		bannerId: {
-			type: 'string',
-			minLength: 1,
-		},
-	},
-	required: ['bannerId'],
-	additionalProperties: false,
-};
-
-export const isBannersDismissProps = ajv.compile<BannersDismiss>(BannersDismissSchema);
-
 export type BannersEndpoints = {
 	'/v1/banners': {
-		GET: (params: Banners) => {
+		GET: (params: { platform: BannerPlatform }) => {
 			banners: IBanner[];
 		};
 	};
 
 	'/v1/banners.dismiss': {
-		POST: (params: BannersDismiss) => void;
+		POST: (params: { bannerId: string }) => void;
 	};
 };
